+++ src/index.ts
@@ -1,7 +1,4 @@
-/*
-export {
-    InterpolatorConstant,
-    InterpolatorLinear,
-    InterpolatorSpline,
-} from "./interpolate";
-*/
+export {InterpolatorBase} from "./base";
+export {InterpolatorConstant} from "./constant";
+export {InterpolatorLinear} from "./linear";
+export {InterpolatorSpline} from "./spline";
